+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- changed: Use new info server variable for thornode server urls with chain name included
 - fixed: Use unique info server cache for each Thorchain-based plugin
 
 ## 2.11.0 (2024-10-08)